refactor(storage): extract ProfileStoreMixin from ProfileMixin (SQLite + ABC)
First Phase-B extraction of the _profiles.py decomposition. Move the
ProfileStore bucket (20 public methods + PS-only privates) verbatim out of
ProfileMixin into a new ProfileStoreMixin, mirroring the merged playbook
decomposition (#262/#491).

- sqlite_storage: new profiles/_profile_store.py with ProfileStoreMixin;
  _emit_hard_delete_profile + _escape_like_pattern move with it (PS-only);
  _build_tags_sql stays module-level in _profiles.py (shared with search) and
  is imported. Residual ProfileMixin keeps interaction + search methods.
- storage_base (ABC): matching ProfileStoreMixin with the abstract surface.
- Compose ProfileStoreMixin alongside the shrinking ProfileMixin in both
  __init__ base lists (+ __all__).
- Bodies are byte-identical; the mutation -> rowcount-guarded event -> commit
  -> self-committing FTS/vec ordering is preserved unchanged.
- Repoint the Task-1 atomicity characterization test's _append_event_stmt
  patch target to the new profiles._profile_store module.
