+++ reflexio/server/services/storage/sqlite_storage/__init__.py
@@ -32,10 +32,12 @@ from .playbook import (
     PlaybookSourceLinkageMixin,
     UserPlaybookStoreMixin,
 )
+from .profiles import ProfileStoreMixin
 
 
 class SQLiteStorage(
     SQLiteAgentRunMixin,
+    ProfileStoreMixin,
     ProfileMixin,
     RequestMixin,
     AgentPlaybookStoreMixin,
