+++ reflexio/server/services/storage/sqlite_storage/profiles/__init__.py
@@ -1,0 +1,5 @@
+from ._profile_store import ProfileStoreMixin
+
+__all__ = [
+    "ProfileStoreMixin",
+]
