+++ reflexio/server/services/storage/storage_base/profiles/__init__.py
@@ -1,0 +1,5 @@
+from ._profile_store import ProfileStoreMixin
+
+__all__ = [
+    "ProfileStoreMixin",
+]
